# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nombre de tu API",
        "version": "1.0.0",
        "description": "Descripción de tu API"
    },
    "servers": [{
        "url": "http://localhost:8080",
        "description": "Servidor de desarrollo"
    }],
    "paths": {
...
        "/api/episodes": {
            "get": {
                "summary": "Obtener todos los episodios",
                "description": "Devuelve todos los episodios.",
                "tags": ["Episodes"],
                "responses": {
                    "200": {
                        "description": "Lista de todos los episodios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Episode"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró"
                    }
                }
            },
            "post": {
                "summary": "Agregar un nuevo episodio",
                "description": "Agrega un nuevo episodio.",
                "tags": ["Episodes"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "img": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "La imagen del episodio (opcional)"
                                    },
                                    "name": {
                                        "type": "string",
                                        "description": "El nombre del episodio"
                                    },
                                    "video": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "El video del episodio"
                                    },
                                    "fullname": {
                                        "type": "string",
                                        "description": "El nombre completo del episodio"
                                    },
                                    "serieId": {
                                        "type": "integer",
                                        "description": "El ID de la serie a la que pertenece el episodio"
                                    },
                                    "episode_number": {
                                        "type": "integer",
                                        "description": "El número del episodio"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Episodio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Episode"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Fallo en la validacion"
                    },
                    "409": {
                        "description": "Clave Primaria duplicada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/episodes/name/{name}": {
            "get": {
                "summary": "Buscar un episodio por nombre",
                "description": "Busca un episodio por su nombre.",
                "tags": ["Episodes"],
                "parameters": [{
                    "name": "name",
                    "in": "path",
                    "required": true,
                    "description": "Nombre del episodio a buscar",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Episodio encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Episode"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró el episodio"
                    }
                }
            }
        },
        "/api/episodes/episodes/{id}": {
            "get": {
                "summary": "Obtener todos los episodios de una serie",
                "description": "Devuelve todos los episodios de una serie específica.",
                "tags": ["Episodes"],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "Identificador de la serie",
                    "schema": {
                        "type": "integer"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Lista de episodios de la serie encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Episode"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró la serie"
                    }
                }
            }
        },
        "/api/episodes/{id}": {
            "put": {
                "summary": "Actualizar un episodio",
                "description": "Actualiza un episodio existente.",
                "tags": ["Episodes"],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID del episodio a actualizar",
                    "schema": {
                        "type": "integer"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "img": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "La nueva imagen del episodio (opcional)"
                                    },
                                    "name": {
                                        "type": "string",
                                        "description": "El nuevo nombre del episodio"
                                    },
                                    "video": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "El nuevo video del episodio (opcional)"
                                    },
                                    "fullname": {
                                        "type": "string",
                                        "description": "El nuevo nombre completo del episodio"
                                    },
                                    "episode_number": {
                                        "type": "integer",
                                        "description": "El nuevo número del episodio"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Episodio actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Episode"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Fallo en la validacion"
                    },
                    "409": {
                        "description": "Clave Primaria duplicada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar un episodio",
                "description": "Elimina un episodio existente.",
                "tags": ["Episodes"],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID del episodio a eliminar",
                    "schema": {
                        "type": "integer"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Episodio eliminado exitosamente"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Serie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "El ID de la serie"
                    },
                    "name": {
                        "type": "string",
                        "description": "El nombre de la serie"
                    },
                    "descr": {
                        "type": "string",
                        "description": "La descripción de la serie"
                    },
                    "imgPath": {
                        "type": "string",
                        "description": "El nombre del archivo de imagen de la serie"
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date",
                        "description": "La fecha de lanzamiento de la serie"
                    },
                    "finishDate": {
                        "type": "string",
                        "format": "date",
                        "description": "La fecha de finalización de la serie"
                    },
                    "search": {
                        "type": "string",
                        "description": "Texto de búsqueda relacionado con la serie"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        },
                        "description": "El conjunto de categorías asociadas a la serie"
                    }
                }
            },
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "El ID de la categoría"
                    },
                    "name": {
                        "type": "string",
                        "description": "El nombre de la categoría"
                    },
                    "search": {
                        "type": "string",
                        "description": "Texto de búsqueda relacionado con la categoría"
                    }
                }
            },
...
            "Episode": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "El identificador único del episodio."
                    },
                    "name": {
                        "type": "string",
                        "description": "El nombre del episodio."
                    },
                    "fullname": {
                        "type": "string",
                        "description": "El nombre completo del episodio."
                    },
                    "imgPath": {
                        "type": "string",
                        "description": "La ruta de la imagen del episodio."
                    },
                    "episodeNumber": {
                        "type": "integer",
                        "description": "El número de episodio."
                    },
                    "videoPath": {
                        "type": "string",
                        "description": "La ruta del video del episodio."
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date",
                        "description": "La fecha de lanzamiento del episodio (en formato 'YYYY-MM-DD')."
                    },
                    "serie": {
                        "$ref": "#/components/schemas/Serie",
                        "description": "La serie a la que pertenece este episodio."
                    }
                },
                "required": ["name", "fullname", "episodeNumber", "serie"],
                "title": "Episode"
            },
...
        }
    }
}